BUG: In PointHandleRepr2D::BuildRepresentation, check also renderer MTime
Checking for Renderer MTime allows to make sure the focal point
is updated when the size of the renderer (viewport) is changed.

For example, in a lightbox settings, when additional row or column
are added, the size of all renderers is updated. Since the location of
the 2D annotation is managed in the 2D overlay, it's important to make
sure the location of the representation is changed.