ENH: Add TestHandleWidget2DWithMultipleRenderer
Six renderers are added to a vtkRenderWindow.

Each one of this renderer display a sphere.

A vtkHandleWidget is also added to the 1st
renderer (bottom-left corner).

Keys from 1 to 6 allows to change the
current renderer associated with the considered
handle widget.

After switching the renderer associated with
the widget, its location within each renderer is
expected to be the same